requirements.txt
jiwer==3.0.3
librosa==0.10.0
numpy==1.23.5
openai_whisper==20231117
pymcd==0.2.1
pysndfx==0.3.6
soundfile==0.12.1
speechbrain==1.0.0
torch==2.2.2+cu121
torch_stoi==0.2.1
torchaudio==2.2.2
tqdm==4.64.1
librosa==0.9.2
matplotlib
numpy
numba
phonemizer
scipy
tensorboard
Unidecode
amfm_decompy
jieba
transformers
pypinyin
cn2an
gradio==3.50.2
av
mecab-python3
loguru
unidic-lite
cmudict
fugashi
num2words
PyYAML
requests
pyopenjtalk-prebuilt
jaconv
psutil
GPUtil
vector_quantize_pytorch
g2p_en
sentencepiece
pykakasi
langid
WeTextProcessing>=0.1.10; platform_system != "Windows"
